internal/rest/client: Move public facing API funcs to internal and implement Client interface
